python/langchain0dot2dot11__cpython3dot9dot6/pyproject.toml
[project]
name = "langchain0dot2dot11-cpython3dot9dot6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.6"
dependencies = [
    "datadog-api-client==2.25.0",
    "duckduckgo-search==6.2.13",
    "fastapi==0.116.1",
    "httpx==0.28.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.67",
    "langchain-openai==0.2.11",
    "python-dotenv==1.0.1",
    "uvicorn==0.24.0",
]

[dependency-groups]
dev = [
    "pytest==7.4.4",
]
